+++ server/conf/db.js
@@ -1,0 +1,4 @@
+const MONGODB_URI =
+  process.env.MONGODB_URI || "mongodb://localhost/edwardnunez";
+
+module.exports = MONGODB_URI;