+++ server/models/tags.js
@@ -1,0 +1,22 @@
+const { Schema, model } = require("mongoose");
+
+const tagSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: true,
+    },
+  },
+  {
+    toJSON: {
+      transform: function (doc, ret) {
+        ret.id = ret._id;
+        delete ret._id;
+      },
+    },
+  }
+);
+
+const tags = model("tags", tagSchema);
+
+module.exports = tags;