+++ src/components/elements/Input.js
@@ -1,39 +1,25 @@
-import React, { useState } from "react";
+import { useState } from 'react';
+import PropTypes from 'prop-types';
 
-import InputLabel from "./InputLabel";
+import InputLabel from './InputLabel';
 
-/**
- * TODO
- */
-export default function Input({
-  elemType,
-  type,
-  id,
-  placeholder,
-  rows,
-  required,
-  label,
-  index,
-  valid,
-  onChange,
-  onClick,
-}) {
-  const Element = elemType === "textarea" ? "textarea" : "input";
+function Input({ elemType, type, id, placeholder, rows, required, label, index, valid, updateInputState }) {
+  const Element = elemType === 'textarea' ? 'textarea' : 'input';
   const [touched, setTouched] = useState(false);
 
   const handleOnChange = (evt) => {
-    onChange(evt, index);
+    updateInputState(evt, index);
   };
 
   const handleOnClick = (evt) => {
     setTouched(true);
 
-    onClick(evt, index, true);
+    updateInputState(evt, index, true);
   };
 
   const isInvalid = () => {
     if (required && touched) {
-      return !valid ? true : false;
+      return !valid;
     }
 
     return false;
@@ -41,14 +27,14 @@ export default function Input({
 
   return (
     <>
-      {label && <InputLabel id={id}>{label}</InputLabel>}
+      {label && <InputLabel id={id} label={label} />}
       <Element
-        type={elemType !== "textarea" ? type : null}
-        className={`form-control ${isInvalid() ? "is-invalid" : ""}`}
+        type={elemType !== 'textarea' ? type : null}
+        className={`form-control ${isInvalid() ? 'is-invalid' : ''}`}
         id={id}
         name={id}
         placeholder={placeholder}
-        rows={elemType === "textarea" ? (rows ? rows : 3) : null}
+        rows={elemType === 'textarea' ? rows || 3 : null}
         required={required}
         onChange={handleOnChange}
         onClick={handleOnClick}
@@ -57,3 +43,25 @@ export default function Input({
     </>
   );
 }
+
+Input.propTypes = {
+  elemType: PropTypes.string.isRequired,
+  type: PropTypes.string,
+  id: PropTypes.string.isRequired,
+  placeholder: PropTypes.string.isRequired,
+  rows: PropTypes.number,
+  required: PropTypes.bool,
+  label: PropTypes.string.isRequired,
+  index: PropTypes.number.isRequired,
+  valid: PropTypes.bool,
+  updateInputState: PropTypes.func.isRequired,
+};
+
+Input.defaultProps = {
+  type: 'text',
+  rows: 0,
+  valid: null,
+  required: false,
+};
+
+export default Input;