+++ src/components/elements/InputFeedback.js
@@ -1,18 +1,15 @@
-import React from "react";
+import PropTypes from 'prop-types';
 
-export default function InputFeedback({ valid, touched, errorMessage }) {
+function InputFeedback({ valid, touched, errorMessage }) {
   return (
     <>
-      <div
-        className="valid-feedback"
-        style={{ display: valid && touched ? "block" : "none" }}
-      >
+      <div className="valid-feedback" style={{ display: valid && touched ? 'block' : 'none' }}>
         Looks good!
       </div>
       <div
         className="invalid-feedback"
         style={{
-          display: !valid && touched ? "block" : "none",
+          display: !valid && touched ? 'block' : 'none',
         }}
       >
         {errorMessage}
@@ -20,3 +17,15 @@ export default function InputFeedback({ valid, touched, errorMessage }) {
     </>
   );
 }
+
+InputFeedback.propTypes = {
+  valid: PropTypes.bool.isRequired,
+  touched: PropTypes.bool.isRequired,
+  errorMessage: PropTypes.string,
+};
+
+InputFeedback.defaultProps = {
+  errorMessage: null,
+};
+
+export default InputFeedback;