+++ src/components/elements/InputLabel.js
@@ -1,9 +1,16 @@
-import React from "react";
+import PropTypes from 'prop-types';
 
-export default function InputLabel({ id, children }) {
+function InputLabel({ id, label }) {
   return (
     <label htmlFor={id} className="form-label">
-      {children}
+      {label}
     </label>
   );
 }
+
+InputLabel.propTypes = {
+  id: PropTypes.string.isRequired,
+  label: PropTypes.string.isRequired,
+};
+
+export default InputLabel;