+++ src/components/elements/ReCaptcha.js
@@ -1,27 +1,21 @@
-import React from "react";
+import PropTypes from 'prop-types';
 
-import ReCAPTCHA from "react-google-recaptcha";
+import ReCAPTCHA from 'react-google-recaptcha';
 
-/**
- * Simple ReCAPTCHA to prevent bots and spammers from abusing functionality.
- *
- * @param {Object} props.onChange Used to listen for changes of the reCAPTCHA validity.
- */
-export default function ReCaptcha(props) {
-  function handleValidChange(valid) {
+function ReCaptcha({ updateReCaptchaState }) {
+  const handleOnChange = (valid) => {
     if (valid !== null) {
-      props.onChange(true); // True if value has valid captcha response stored
+      updateReCaptchaState(true); // True if value has valid captcha response stored
     } else {
-      props.onChange(false); // False if value is null for no valid captcha response
+      updateReCaptchaState(false); // False if value is null for no valid captcha response
     }
-  }
+  };
 
-  return (
-    <>
-      <ReCAPTCHA
-        sitekey="6LeFmQkTAAAAAIwpph9vOJU7yCZQ2mAWb25B-VET"
-        onChange={handleValidChange}
-      />
-    </>
-  );
+  return <ReCAPTCHA sitekey="6LeFmQkTAAAAAIwpph9vOJU7yCZQ2mAWb25B-VET" onChange={handleOnChange} />;
 }
+
+ReCaptcha.propTypes = {
+  updateReCaptchaState: PropTypes.func.isRequired,
+};
+
+export default ReCaptcha;