+++ src/components/sections/ContactForm.js
@@ -1,168 +1,230 @@
-import React, { useState } from "react";
-import emailjs from "emailjs-com";
+import { useEffect, useState } from 'react';
+import emailjs from 'emailjs-com';
+import { toast } from 'react-toastify';
 
-import Input from "../elements/Input";
-import InputFeedback from "../elements/InputFeedback";
-import ReCaptcha from "../elements/ReCaptcha";
+import Input from '../elements/Input';
+import InputFeedback from '../elements/InputFeedback';
+import ReCaptcha from '../elements/ReCaptcha';
 
 export default function ContactForm() {
   const [formData, setFormData] = useState({
     inputs: [
       {
-        elemType: "input",
-        type: "text",
-        id: "name",
+        elemType: 'input',
+        type: 'text',
+        id: 'name',
         required: true,
-        placeholder: "Name",
-        label: "Name",
+        placeholder: 'Name',
+        label: 'Name',
         valid: false,
-        errorMessage: "Please fill out this field.",
+        errorMessage: 'Please fill out this field.',
         touched: false,
-        value: "",
+        value: '',
       },
       {
-        elemType: "input",
-        type: "email",
-        id: "email",
+        elemType: 'input',
+        type: 'email',
+        id: 'email',
         required: true,
-        placeholder: "E-mail",
-        label: "E-mail",
+        placeholder: 'E-mail',
+        label: 'E-mail',
         valid: false,
         touched: false,
-        errorMessage: "Please fill out this field.",
-        value: "",
+        errorMessage: 'Please fill out this field.',
+        value: '',
       },
       {
-        elemType: "input",
-        type: "text",
-        id: "company",
+        elemType: 'input',
+        type: 'text',
+        id: 'company',
         required: false,
-        placeholder: "Company",
-        label: "Company (optional)",
+        placeholder: 'Company',
+        label: 'Company (optional)',
         valid: true,
         touched: false,
-        value: "",
+        value: '',
       },
       {
-        elemType: "textarea",
-        id: "message",
+        elemType: 'textarea',
+        id: 'message',
         required: true,
-        placeholder: "Message",
-        label: "Message",
+        placeholder: 'Message',
+        label: 'Message',
         valid: false,
-        errorMessage: "Please fill out this field.",
+        errorMessage: 'Please fill out this field.',
         touched: false,
         rows: 3,
-        value: "",
+        value: '',
       },
     ],
-    reCaptcha: { touched: false, valid: false },
-    allFieldsValid: false,
     emailSent: false,
   });
 
-  const handleReCAPTCHA = (valid) => {
-    let newFormData = { ...formData };
-    newFormData.reCaptcha.valid = valid;
-    newFormData.reCaptcha.touched = true;
+  const [reCaptchaData, setReCaptchaData] = useState({
+    touched: false,
+    valid: false,
+  });
 
-    setFormData(newFormData);
+  const [formValid, setFormValid] = useState(false);
+
+  useEffect(() => {
+    const formValidation = () => {
+      let isFormValid = true;
+
+      formData.inputs.forEach((input) => {
+        if (input.valid === false) {
+          isFormValid = false;
+        }
+      });
+
+      if (!reCaptchaData.valid) {
+        isFormValid = false;
+      }
+
+      return isFormValid;
+    };
+
+    setFormValid(formValidation());
+  }, [formData, reCaptchaData]);
+
+  const updateReCaptchaState = (valid) => {
+    const newData = { valid, touched: true };
+
+    setReCaptchaData(newData);
   };
 
-  const handleInputEvent = (evt, index, touched) => {
-    let newFormData = { ...formData };
+  const updateInputState = (evt, index, touched) => {
+    const newFormData = { ...formData };
+    newFormData.inputs[index].errorMessage = evt.target.validationMessage;
+    newFormData.inputs[index].valid = evt.target.validity.valid;
+    newFormData.inputs[index].value = evt.target.value;
 
     if (touched) {
       newFormData.inputs[index].touched = touched;
     }
 
-    newFormData.inputs[index].errorMessage = evt.target.validationMessage;
-    newFormData.inputs[index].valid = evt.target.validity.valid;
-    newFormData.inputs[index].value = evt.target.value;
-
-    setFormData(newFormData);
-  };
-
-  const handleValidity = () => {
-    let allFieldsValid = true;
-    const newFormData = { ...formData };
-
-    newFormData.inputs.forEach((input) => {
-      if (input.valid === false) {
-        allFieldsValid = false;
-      }
-
-      input.touched = true;
-    });
-
-    newFormData.allFieldsValid = allFieldsValid;
-    setFormData(newFormData);
+    setFormData((oldFormData) => ({ ...oldFormData, ...newFormData }));
   };
 
   const handleSubmit = (evt) => {
     evt.preventDefault(); // Stops page from reloading when user presses Submit.
-    handleValidity();
 
-    console.log(formData.reCaptcha.valid, formData.allFieldsValid);
-    if (formData.reCaptcha.valid && formData.allFieldsValid) {
-      emailjs
-        .sendForm(
-          "edwardnnz_service",
-          "contact_form",
-          evt.target,
-          "user_907s5tO5hxVUSKyRsGi0U"
-        )
-        .then(
-          (result) => {
-            console.log(result.text);
-            setFormData({ ...formData, emailSent: true });
-          },
-          (error) => {
-            console.log(error.text);
-          }
-        );
+    if (formValid) {
+      emailjs.sendForm('service_ccws3ss', 'contact_form', evt.target, 'user_907s5tO5hxVUSKyRsGi0U').then(
+        (result) => {
+          // Succesfull
+          toast.success('Email has been sent, thank you!', {
+            position: 'bottom-right',
+            autoClose: 5000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: 'light',
+          });
+
+          setFormData({ ...formData, emailSent: true });
+        },
+        (error) => {
+          // Failed
+          toast.error('Failed to send email, please try again.', {
+            position: 'bottom-right',
+            autoClose: 5000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: 'light',
+          });
+        }
+      );
     }
   };
 
-  const renderInputs = () =>
-    formData.inputs.map((input, index) => {
-      return (
-        <div className="mb-3" key={index}>
-          <Input
-            {...input}
-            index={index}
-            onChange={handleInputEvent}
-            onClick={handleInputEvent}
-          />
-          <InputFeedback
-            valid={input.valid}
-            touched={input.touched}
-            errorMessage={input.errorMessage}
-          />
-        </div>
-      );
-    });
-
   return (
     <>
-      <div style={{ display: `${formData.emailSent ? "block" : "none"}` }}>
-        Email sent!
-      </div>
-      <form
-        onSubmit={handleSubmit}
-        style={{ display: `${formData.emailSent ? "none" : "block"}` }}
-        noValidate
-      >
-        {renderInputs()}
+      <div style={{ display: `${formData.emailSent ? 'block' : 'none'}` }}>Email sent!</div>
+      <form onSubmit={handleSubmit} style={{ display: `${formData.emailSent ? 'none' : 'block'}` }} noValidate>
+        <div className="mb-3" key={formData.inputs[0].id}>
+          <Input
+            elemType={formData.inputs[0].elemType}
+            type={formData.inputs[0].type}
+            id={formData.inputs[0].id}
+            placeholder={formData.inputs[0].placeholder}
+            rows={formData.inputs[0].rows}
+            required={formData.inputs[0].required}
+            label={formData.inputs[0].label}
+            index={0}
+            updateInputState={updateInputState}
+          />
+          <InputFeedback
+            valid={formData.inputs[0].valid}
+            touched={formData.inputs[0].touched}
+            errorMessage={formData.inputs[0].errorMessage}
+          />
+        </div>
+        <div className="mb-3" key={formData.inputs[1].id}>
+          <Input
+            elemType={formData.inputs[1].elemType}
+            type={formData.inputs[1].type}
+            id={formData.inputs[1].id}
+            placeholder={formData.inputs[1].placeholder}
+            rows={formData.inputs[1].rows}
+            required={formData.inputs[1].required}
+            label={formData.inputs[1].label}
+            index={1}
+            updateInputState={updateInputState}
+          />
+          <InputFeedback
+            valid={formData.inputs[1].valid}
+            touched={formData.inputs[1].touched}
+            errorMessage={formData.inputs[1].errorMessage}
+          />
+        </div>
+        <div className="mb-3" key={formData.inputs[2].id}>
+          <Input
+            elemType={formData.inputs[2].elemType}
+            type={formData.inputs[2].type}
+            id={formData.inputs[2].id}
+            placeholder={formData.inputs[2].placeholder}
+            rows={formData.inputs[2].rows}
+            required={formData.inputs[2].required}
+            label={formData.inputs[2].label}
+            index={2}
+            updateInputState={updateInputState}
+          />
+          <InputFeedback
+            valid={formData.inputs[2].valid}
+            touched={formData.inputs[2].touched}
+            errorMessage={formData.inputs[2].errorMessage}
+          />
+        </div>
+        <div className="mb-3" key={formData.inputs[3].id}>
+          <Input
+            elemType={formData.inputs[3].elemType}
+            type={formData.inputs[3].type}
+            id={formData.inputs[3].id}
+            placeholder={formData.inputs[3].placeholder}
+            rows={formData.inputs[3].rows}
+            required={formData.inputs[3].required}
+            label={formData.inputs[3].label}
+            index={3}
+            updateInputState={updateInputState}
+          />
+          <InputFeedback
+            valid={formData.inputs[3].valid}
+            touched={formData.inputs[3].touched}
+            errorMessage={formData.inputs[3].errorMessage}
+          />
+        </div>
         <div className="mb-3">
-          <ReCaptcha onChange={handleReCAPTCHA} />
+          <ReCaptcha updateReCaptchaState={updateReCaptchaState} />
           <div
             className="invalid-feedback"
             style={{
-              display:
-                !formData.reCaptcha.valid && formData.reCaptcha.touched
-                  ? "block"
-                  : "none",
+              display: !reCaptchaData.valid && reCaptchaData.touched ? 'block' : 'none',
             }}
           >
             Please complete ReCAPTCHA