+++ src/views/NoMatch.js
@@ -1,9 +1,3 @@
-import React from "react";
-
 export default function NoMatch() {
-  return (
-    <>
-      <h1>404 not found</h1>
-    </>
-  );
+  return <h1>404 not found</h1>;
 }